+++ tests/extractors/test_xml.py
@@ -748,3 +748,15 @@ def test_domain_filter():
                 </html>'''.encode('utf-8')
     messages = list(xml_extractor('filename', _options(domain='other')))
     assert len(messages) == 0
+
+
+@pytest.mark.usefixtures('fake_source')
+def test_domain_filter_for_attribute():
+    global source
+    source = u'''\
+                <html xmlns:i18n="http://xml.zope.org/namespaces/i18n"
+                      i18n:domain="lingua">
+                  <dummy title="Dummy text" i18n:attributes="title"></dummy>
+                </html>'''.encode('utf-8')
+    messages = list(xml_extractor('filename', _options(domain='other')))
+    assert len(messages) == 0