+++ changes.rst
@@ -4,6 +4,9 @@ Changelog
 3.12 - October 5, 2015
 ----------------------
 
+- Fix HTML extractor to Apply domain filter when processing attributes. This
+  fixes `issue 66 <https://github.com/wichert/lingua/issues/66`_
+
 - Fix line number reporting when processing instructions are used.
   `Patch #64 <https://github.com/wichert/lingua/pull/64>`_ from
   Florian Schulze.